Make the casting a bit less sneaky.

+++ src/server.cpp
@@ -146,7 +146,7 @@ int main()
             case ENET_EVENT_TYPE_DISCONNECT:
             {
                 logInfo("Disconnect from %x:%u\n", netEvent.peer->address.host, netEvent.peer->address.port);
-                UserIdentifier oldUserId = (UserIdentifier)(((int64_t)netEvent.peer->data) & 0xFF);
+                UserIdentifier oldUserId = (UserIdentifier)(((intptr_t)netEvent.peer->data) & 0xFF);
                 if(oldUserId == 0)
                     break;
